Initial complete functioning frontend/UI
This adds a user interface in the extension's options page.
The page loads and lists all the containers with their names,
their icons and icons colors, as well as an input box that is
intended for inputting containers' default URLs.
Currently, it's a dumb box; inputting will not save anything.
Also, UI needs improvment in being responsive and accessible.