Prevent reloading option page when entering new input
Previously, the HTML option page reloads every time the user
presses enter to save input. This causes a white flash which
is especially apparent when having a dark theme. The flash is
still present on initial page load. This commit at least
prevents that from happening while the user is on the page
every time they hit enter to save the input.
This might however introduce an issue, where now there is no
indication when the input has been received. So, additional
interface is most likely needed to make submission and saving
the input obvious, and hopefully intiutive.