Add frontend responsive design.
The option page now alignes the identity info i.e. icon and name
as one item to the left and the input boxs to the right with a
flexible empty space between them. When the page shrinks in size
the space between also shrinks with fixed size for the input box
and a minimum size for the identity info. The identity name is
also flexible. If it is too long it will be truncated adding
ellipses. This name truncation occurs both when the page shrinks
(i.e. even if the name is short) and when the name itself is
long exceeding the white space that is between the info and the
input box. The name still has a minimum size where all default
firefox names fit without truncation. The icon also has a fixed
so all icons aesthetically occupy the same area.