Add a save button
Add a save button next to the url box so user can click it and
url is sent to store in backend. When the page loads, the button
shows "Saved". When the user types something in the box, the
button changes from "Saved" to "Save". Clicking on it saves
the url and changes the button back to "Saved". Currently both
"Save" and "Saved" buttons have the same style but the script
toggles the class attribute between "save-btn" and "saved-btn"
so to enable css styling.