Make options page more accessible
More likely there is more to be done as I am still learning this
But for now I added several aria-labels. I also did a blind test
with a screen reader and made some tweeks. The hr separator is
specially helpful with the screen reader as it makes it easier
to distinguish one container row from another. As I learn more,
I will try to improve this experience.