Add garbage collection in backend
Backend now listens for events of deleting containers.
When a container is deleted, the backend deletes its
storage input from storage.
A possible scenario where this is bypassed is when the
delete operation fails, at which currently it does not
retry. So, that's possible TODO, though this is unlikely.
In that case, we could try delayed retrys or checking
existing container against those in storage on next
extension launch.