Add cookies permission to access tab.cookieSoreId, and up the version
cookies permission is needed to access tab.cookieStoreId which
is critical, to associate a tab with its container. That was not
added before as tab.cookieSoreId could be accessed on Firefox
Developer Edition where the extension was first tested but Firefox
does need to enable this permission. From the docs:
https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/tabs/Tab

"cookieStoreId Optional
    string. The cookie store of the tab. If different tabs can have
    different cookie stores (for example, to support contextual identity),
    you can pass this as the storeId option into various methods of the
    cookies API, to set and get cookies associated with this tab's cookie store.
    Only present if the extension has the "cookies" permission."